Start using the new columns a little

+++ src/isopod/sender2.py
@@ -25,13 +25,13 @@ class Sender(Controller):
         if (result := self._reconcile_with_rsync()) is not None:
             return result
 
-        if (path := self._get_next_path()) is None:
+        if (disc := self._get_next_disc()) is None:
             log.info("Waiting for next sendable disc")
             return Reconciled()
 
-        args = ["rsync", "--partial", path, f"{self.target_base}/{path}"]
+        args = ["rsync", "--partial", disc.path, f"{self.target_base}/{disc.path}"]
         self._rsync = Popen(args, stdin=DEVNULL, stdout=DEVNULL, stderr=DEVNULL)
-        self._current_path = path
+        self._current_disc = disc
         self._poll_on_finish(self._rsync)
         log.info("Started: %s", shlex.join(args))
         return Reconciled()
@@ -49,11 +49,8 @@ class Sender(Controller):
         if (returncode := self._rsync.poll()) is None:
             return Reconciled()
 
-        path = self._current_path
-        self._rsync = None
-        self._current_path = None
         if returncode == 0:
-            self._finalize_rsync_success(path)
+            self._finalize_rsync_success()
         else:
             # TODO: Improved backoff strategy: handle the case of a
             # single ISO being unsendable (e.g. unexpected deletion) and
@@ -63,18 +60,23 @@ class Sender(Controller):
 
         return None
 
-    def _finalize_rsync_success(self, path):
+    def _finalize_rsync_success(self):
         with db.Session() as session:
-            stmt = select(db.Disc).filter_by(path=path)
-            disc = session.execute(stmt).scalar_one()
-            disc.status = db.DiscStatus.COMPLETE
-            session.commit()
-            log.info("Finished sending %s", path)
+            if (disc := self._current_disc) is None:
+                raise TypeError("missing current disc")
 
-            isopod.force_unlink(path)
+            self._rsync = None
+            self._current_disc = None
+
+            disc.status = db.DiscStatus.COMPLETE
+            session.merge(disc)
+            session.commit()
+            log.info("Finished sending %s", disc.path)
+
+            isopod.force_unlink(disc.path)
             session.delete(disc)
             session.commit()
-            log.info("Cleaned up %s", path)
+            log.info("Cleaned up %s", disc.path)
 
     def _poll_on_finish(self, proc: Popen):
         def wait_and_poll():
@@ -83,9 +85,11 @@ class Sender(Controller):
 
         Thread(target=wait_and_poll, daemon=True).start()
 
-    def _get_next_path(self):
+    def _get_next_disc(self):
         with db.Session() as session:
-            stmt = select(db.Disc).filter_by(status=db.DiscStatus.SENDABLE)
-            disc = session.execute(stmt).scalars().first()
-            if disc is not None:
-                return disc.path
+            stmt = (
+                select(db.Disc)
+                .filter_by(status=db.DiscStatus.SENDABLE)
+                .order_by(db.Disc.send_attempts, db.Disc.last_send_failure)
+            )
+            return session.execute(stmt).scalars().first()
